requirements.txt
boto3>=1.28.0
flask>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
